+++ Interfaces/IWorkflowBus.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 using OptimaJet.Workflow.Core.Runtime;
 
 namespace OptimaJet.Workflow.Core.Bus
@@ -23,29 +25,34 @@ namespace OptimaJet.Workflow.Core.Bus
         /// Queue execution with the list of <see cref="ExecutionRequestParameters"/>
         /// </summary>
         /// <param name="requestParameters">List of <see cref="ExecutionRequestParameters"/></param>
+        /// <param name="token">Cancellation token</param>
         /// <param name="notFireExecutionComplete">If true - the Bus must execute the Request without firing ExecutionComplete</param>
-        void QueueExecution(IEnumerable<ExecutionRequestParameters> requestParameters, bool notFireExecutionComplete = false);
+        /// <returns>True if some activity and actions were executed</returns>
+        Task<bool> QueueExecution(IEnumerable<ExecutionRequestParameters> requestParameters, CancellationToken token, bool notFireExecutionComplete = false);
 
         /// <summary>
         /// Queue execution with the <see cref="ExecutionRequestParameters"/>
         /// </summary>
         /// <param name="requestParameters">Instance of <see cref="ExecutionRequestParameters"/></param>
+        /// /// <param name="token">Cancellation token</param>
         /// <param name="notFireExecutionComplete">If true - the Bus must execute the Request without firing ExecutionComplete</param>
-        void QueueExecution(ExecutionRequestParameters requestParameters, bool notFireExecutionComplete = false);
+        /// <returns>True if some activity and actions were executed</returns>
+        Task<bool> QueueExecution(ExecutionRequestParameters requestParameters, CancellationToken token, bool notFireExecutionComplete = false);
 
         /// <summary>
         /// Raises after the execution was complete
         /// </summary>
-        event EventHandler<ExecutionResponseEventArgs> ExecutionComplete;
+        // event EventHandler<ExecutionResponseEventArgs> ExecutionComplete;
+        event Func<IWorkflowBus,ExecutionResponseEventArgs, Task> ExecutionComplete;
 
         /// <summary>
         /// Raises before execution of choosen activity
         /// </summary>
         event EventHandler<BeforeActivityExecutionEventArgs> BeforeExecution;
 
-        /// <summary>
-        /// Returns true if the bus supports asynchronous model of execution
-        /// </summary>
-        bool IsAsync { get; }
+        ///// <summary>
+        ///// Returns true if the bus supports asynchronous model of execution
+        ///// </summary>
+        //bool IsAsync { get; }
     }
 }