+++ Provider for MySQL/Scripts/CreatePersistenceObjects.sql
@@ -1,7 +1,7 @@
 /*
 Company: OptimaJet
 Project: WorkflowEngine.NET Provider for MySQL
-Version: 2.1
+Version: 2.2
 File: CreatePersistenceObjects.sql
 */