+++ Samples/MSSQL/WF.Sample.Business/Workflow/WorkflowActions.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 using OptimaJet.Workflow.Core.Model;
 using OptimaJet.Workflow.Core.Runtime;
 
@@ -66,6 +68,11 @@ namespace WF.Sample.Business.Workflow
             throw new NotImplementedException(string.Format("Action with name {0} not implemented", name));
         }
 
+        public async Task ExecuteActionAsync(string name, ProcessInstance processInstance, WorkflowRuntime runtime, string actionParameter, CancellationToken token)
+        {
+            throw new NotImplementedException();
+        }
+
         public bool ExecuteCondition(string name, ProcessInstance processInstance, WorkflowRuntime runtime, string actionParameter)
         {
             if (_conditions.ContainsKey(name))
@@ -74,6 +81,21 @@ namespace WF.Sample.Business.Workflow
             }
 
             throw new NotImplementedException(string.Format("Action condition with name {0} not implemented", name));
+        }
+
+        public async Task<bool> ExecuteConditionAsync(string name, ProcessInstance processInstance, WorkflowRuntime runtime, string actionParameter, CancellationToken token)
+        {
+            throw new NotImplementedException();
+        }
+
+        public bool IsActionAsync(string name)
+        {
+            return false;
+        }
+
+        public bool IsConditionAsync(string name)
+        {
+            return false;
         }
 
         public List<string> GetActions()